Removed VirtualVec. Using the master vec to ensure layout is preserved in derived data. So FoldingColumn does not take empty collections of columns anymore.
Also, pulled out the stuff that people don't like in TestUtil, creating a separate class for that, TestBase.

WIP.
