+++ h2o-core/src/test/java/water/TestUtil.java
@@ -1,7 +1,10 @@
 package water;
 
 import org.apache.commons.io.FileUtils;
-import org.junit.*;
+import org.junit.AfterClass;
+import org.junit.Assert;
+import org.junit.Ignore;
+import org.junit.Rule;
 import org.junit.rules.TestRule;
 import org.junit.runner.Description;
 import org.junit.runners.model.Statement;
@@ -19,11 +22,6 @@ import java.io.File;
 import java.io.IOException;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.HashSet;
-import java.util.Set;
-import java.util.function.Consumer;
 import java.util.*;
 
 import static org.junit.Assert.assertEquals;
@@ -32,11 +30,6 @@ import static org.junit.Assert.assertTrue;
 
 @Ignore("Support for tests, but no actual tests here")
 public class TestUtil extends Iced {
-  { // need this because -ea IS NOT ALWAYS set in intellij
-    ClassLoader loader = getClass().getClassLoader();
-    loader.setDefaultAssertionStatus(true);
-  }
-
   public final static boolean JACOCO_ENABLED = Boolean.parseBoolean(System.getProperty("test.jacocoEnabled", "false"));
   private static boolean _stall_called_before = false;
   private static String[] ignoreTestsNames;
@@ -239,7 +232,7 @@ public class TestUtil extends Iced {
     assertTrue("File should exist: " + name, file.exists());
   }
 
-  private static void checkFile(String name, File file) {
+  protected static void checkFile(String name, File file) {
     checkFileEntry(name, file);
     assertTrue("Expected a readable file: " + name, file.canRead());
   }
@@ -257,19 +250,10 @@ public class TestUtil extends Iced {
    *  @return      Frame or NPE */
   public static Frame parse_test_file( String fname ) { return parse_test_file(Key.make(),fname); }
   public static Frame parse_test_file( Key outputKey, String fname) {
-    Vec data = loadFile(fname);
-    return ParseDataset.parse(outputKey, data._key);
-  }
-
-  public static Vec loadFile(String fname) {
-    File f = getFile(fname);
-    return NFSFileVec.make(f);
-  }
-
-  public static File getFile(String fname) {
     File f = find_test_file_static(fname);
     checkFile(fname, f);
-    return f;
+    NFSFileVec nfs = NFSFileVec.make(f);
+    return ParseDataset.parse(outputKey, nfs._key);
   }
 
   protected Frame parse_test_file( Key outputKey, String fname , boolean guessSetup) {
@@ -280,9 +264,11 @@ public class TestUtil extends Iced {
   }
 
   protected Frame parse_test_file( String fname, String na_string, int check_header, byte[] column_types ) {
-    Vec data = loadFile(fname);
+    File f = find_test_file_static(fname);
+    checkFile(fname, f);
+    NFSFileVec nfs = NFSFileVec.make(f);
 
-    Key[] res = {data._key};
+    Key[] res = {nfs._key};
 
     // create new parseSetup in order to store our na_string
     ParseSetup p = ParseSetup.guessSetup(res, new ParseSetup(DefaultParserProviders.GUESS_INFO,(byte) ',',true,
@@ -732,25 +718,4 @@ public class TestUtil extends Iced {
     }
   }
 
-  static Consumer<Vec> dropit = new Consumer<Vec>() {
-    @Override public void accept(Vec v) { v.remove(new Futures()).blockForPending(); }
-  };
-
-  @BeforeClass
-  public static void hi() {
-    stall_till_cloudsize(1);
-  }
-
-  @Before public void enterScope() {
-    Scope.enter();
-  }
-  
-  @After public void bye() { Scope.exit(); }
-
-  protected static Vec willDrop(Vec v) { return Scope.track(v); }
-
-  protected static <T extends Vec.Holder> T willDrop(T vh) {
-    Scope.track(vh.vec());
-    return vh;
-  }
 }
